Support the show_as_available and not_available_label options
When show_as_available is ‘true’, the room will appear as “available” if
no events are booked. This is the existing behaviour.

When ‘false’, the room will appear as ‘not in use’ - this is aimed at 
resources which have a formal booking process (eg. a research lab). A 
custom label must be provided as the ‘not_available_label’ parameter.